* Components/ArticleTable/ArticleTablePlugin.h         * Components/ArticleView/ArticleTextViewPlugin.m         * NS(U)Integer transitions
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2260 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/ArticleView/ArticleTextViewPlugin.m
@@ -130,10 +130,10 @@
  */
 -(BOOL) textView: (NSTextView*) textView
    clickedOnLink: (id) link
-         atIndex: (unsigned) charIndex
+         atIndex: (NSUInteger) charIndex
 {
     BOOL result = NO;
-    NSLog(@"textView:clickedOnLink: %@ atIndex: %d", link, charIndex);
+    NSLog(@"textView:clickedOnLink: %@ atIndex: %"PRIuPTR"", link, charIndex);
     
     if ([link isKindOfClass: [NSURL class]]) {
         result = [[NSWorkspace sharedWorkspace] openURL: (NSURL*)link];
